Adding working version of arbiter for python
This uses the test call of arbiter to setup N tests to interoperate,
based on language and action provided. It sets up unique code in a temp
directory to execute. It runs the code with a timeout, then kills
everything and checks stdout for expected results defined in the code
segment. If everything works out fine it will also return cleanup and
delete the temp directory, if not it will print out the location of it
so you can track down what happened.

Please test by calling:
 arbiter.py -f test -a "python register:Reg/Call Basic 1" -a "python call:Reg/Call Basic 1"
 arbiter.py -f test -a "python subscribe:Pub/Sub Basic" -a "python publish:Pub/Sub Basic"